checkpoint spring 2

+++ routes/web.php
@@ -12,10 +12,10 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\UserProfileController;
 
 
+
 Route::get('/', function () {
     return view('privacy-policy');
 });
-
 
 
 Auth::routes();
@@ -32,24 +32,24 @@ Route::middleware(['auth', 'admin'])->group(function () {
 
 
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home')->middleware('auth');
 
 //Route::resource('users', UserController::class);
-Route::resource('items', ItemController::class);
-Route::resource('reservations', ReservationController::class);
-Route::resource('damage-reports', DamageReportController::class);
+Route::resource('items', ItemController::class)->middleware('auth');
+Route::resource('reservations', ReservationController::class)->middleware('auth');
+Route::resource('damage-reports', DamageReportController::class)->middleware('auth');
 
-Route::get('items', [ItemController::class, 'index'])->name('items.index');
-Route::get('items/create', [ItemController::class, 'create'])->name('items.create');
-Route::post('items', [ItemController::class, 'store'])->name('items.store');
-
-
-Route::resource('reservations', ReservationController::class);
+Route::get('items', [ItemController::class, 'index'])->name('items.index')->middleware('auth');
+Route::get('items/create', [ItemController::class, 'create'])->name('items.create')->middleware('auth');
+Route::post('items', [ItemController::class, 'store'])->name('items.store')->middleware('auth');
 
 
-Route::get('reservations/{reservation_id}/edit', [ReservationController::class, 'edit'])->name('reservations.edit');
-Route::put('reservations/{reservation_id}', [ReservationController::class, 'update'])->name('reservations.update');
-Route::delete('reservations/{reservation_id}', [ReservationController::class, 'destroy'])->name('reservations.destroy');
+Route::resource('reservations', ReservationController::class)->middleware('auth');
+
+
+Route::get('reservations/{reservation_id}/edit', [ReservationController::class, 'edit'])->name('reservations.edit')->middleware('auth');
+Route::put('reservations/{reservation_id}', [ReservationController::class, 'update'])->name('reservations.update')->middleware('auth');
+Route::delete('reservations/{reservation_id}', [ReservationController::class, 'destroy'])->name('reservations.destroy')->middleware('auth');
 
 
 Route::get('/privacy-policy', function () {
@@ -64,9 +64,11 @@ Route::put('/profile', [UserProfileController::class, 'update'])->name('profile.
 
 
 
-Route::post('/make-admin/{user_id}', [AdminController::class, 'makeAdmin'])->name('admin.make');
-Route::post('/admin/remove/{user_id}', [AdminController::class, 'removeAdmin'])->name('admin.remove');
-Route::delete('/admin/delete/{user_id}', [AdminController::class, 'deleteUser'])->name('admin.delete');
+Route::post('/make-admin/{user_id}', [AdminController::class, 'makeAdmin'])->name('admin.make')->middleware('auth');
+Route::post('/admin/remove/{user_id}', [AdminController::class, 'removeAdmin'])->name('admin.remove')->middleware('auth');
+Route::delete('/admin/delete/{user_id}', [AdminController::class, 'deleteUser'])->name('admin.delete')->middleware('auth');
 
 
+
+Route::resource('damage_reports', DamageReportController::class)->middleware('auth');
